[refactor] entity 관계없이 공통으로 사용되는 service 클래스 디렉토리 분리
